mv ternary/layout/* to layout_defaults.js and layout_attributes.js
- like more 'recent' subplot types.
