--- src/plots/ternary/layout/attributes.js
@@ -1,25 +1,0 @@
-/**
-* Copyright 2012-2018, Plotly, Inc.
-* All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE file in the root directory of this source tree.
-*/
-
-'use strict';
-
-
-module.exports = {
-    subplot: {
-        valType: 'subplotid',
-        role: 'info',
-        dflt: 'ternary',
-        editType: 'calc',
-        description: [
-            'Sets a reference between this trace\'s data coordinates and',
-            'a ternary subplot.',
-            'If *ternary* (the default value), the data refer to `layout.ternary`.',
-            'If *ternary2*, the data refer to `layout.ternary2`, and so on.'
-        ].join(' ')
-    }
-};
